requirements.txt
annotated-types==0.7.0
anyio==4.9.0
certifi==2025.6.15
click==8.2.1
colorama==0.4.6
fastapi==0.115.14
gunicorn==23.0.0
h11==0.16.0
httpcore==1.0.9
httptools==0.6.4
httpx==0.28.1
idna==3.10
packaging==25.0
passlib[bcrypt]
pydantic==2.11.7
pydantic_core==2.33.2
python-dotenv==1.1.1
python-jose[cryptography]
python-multipart==0.0.20
PyYAML==6.0.2
sniffio==1.3.1
starlette==0.46.2
sqlalchemy
typing-inspection==0.4.1
typing_extensions==4.14.0
uvicorn==0.35.0
watchfiles==1.1.0
websockets==15.0.1
zipfile36==0.1.3
